+++ words-webifi.js
@@ -28,9 +28,11 @@ function WordsWebifi(webifi) {
   this.description = 'Word patterns, anagrams, word sounds, definitions, synonyms';
 
   this.lex = null;
-  if (!this.initLex()) {
-    this.loadLex();
-  }
+  this.lexParts = 3;
+  this.lexPartsLoaded = 0;
+  this.loadLex('lufz-en-lexicon.js');
+  this.loadLex('lufz-en-lexicon-stems.js');
+  this.loadLex('exet-lexicon.js');
   this.MAX_WORD_CHOICES = webifi.GROUP_SIZE * webifi.MAX_LIST_LEN;
 
   this.webifi.registerAvatar(this.name, this.description, {
@@ -38,6 +40,11 @@ function WordsWebifi(webifi) {
       description: 'Use dictionary dev dot api to look up a word or phrase.',
       prefixes: ['define|definition|definitions', 'look up', 'definition|definitions|meaning of',],
       helpkeys: ['word', 'words', 'phrase',],
+    },
+    'region': {
+      description: 'Change spellings region.',
+      prefixes: ['region|locale|spelling|spell|spellings',],
+      helpkeys: ['region', 'spell', 'spellings', 'spelling', 'locale'],
     },
     'synonyms': {
       description: 'Use dictionary dev dot api to look up synonyms of a word or phrase.',
@@ -68,31 +75,31 @@ function WordsWebifi(webifi) {
 }
 
 WordsWebifi.prototype.initLex = function() {
-  if ((typeof exetLexicon == 'undefined') || (typeof exetLexiconInit== 'undefined') ) {
-    console.log('Lexicon not yet available');
-    return false;
-  }
-  if (this.lex) {
+  if (this.lexPartsLoaded >= this.lexParts) {
+    console.assert(this.lex);
     console.log('Lexicon already initialized');
     return true;
   }
+  this.lexPartsLoaded++;
+  console.log('Lexicon parts loaded: ' + this.lexPartsLoaded + ' of ' +
+              this.lexParts);
+  if (this.lexPartsLoaded < this.lexParts) {
+    return false;
+  }
+  console.assert(!this.lex);
   console.log('Initializing lexicon');
   exetLexiconInit();
   this.lex = exetLexicon;
   return true;
 }
 
-WordsWebifi.prototype.loadLex = function() {
+WordsWebifi.prototype.loadLex = function(scriptFile) {
   const handler = this.initLex.bind(this);
-  const scriptLufz = document.createElement('script');
-  scriptLufz.src = this.webifi.scriptUrlBase + 'lufz-en-lexicon.js';
-  scriptLufz.onload = handler;
-  scriptLufz.onerror = (ev) => {console.log(ev); console.log('error loading script');}
-  const scriptLex = document.createElement('script');
-  scriptLex.src = this.webifi.scriptUrlBase + 'exet-lexicon.js';
-  scriptLex.onload = handler;
-  document.head.append(scriptLufz);
-  document.head.append(scriptLex);
+  const scriptForLufz = document.createElement('script');
+  scriptForLufz.src = this.webifi.scriptUrlBase + scriptFile;
+  scriptForLufz.onload = handler;
+  scriptForLufz.onerror = (ev) => {console.log(ev); console.log('error loading script: ' + scriptForLufz.src);}
+  document.head.append(scriptForLufz);
 }
 
 WordsWebifi.prototype.syndefHandler = function(input, words, commandName,
@@ -192,6 +199,21 @@ WordsWebifi.prototype.handlePattern = function(pattern) {
   this.webifi.output(this.name, 'Here are some matches.', this.webifi.makeGroupedList(matchingWords), false);
 }
 
+WordsWebifi.prototype.handleRegion = function(region) {
+  if (!this.lex) {
+    this.webifi.output(this.name, 'This command is not available as the lexicon file has not been loaded');
+    return;
+  }
+  const regionName = region || 'None';
+  const newRegion = this.lex.preferRegion(region);
+  const newRegionName = newRegion || 'None';
+  if (newRegion != region) {
+    this.webifi.output(this.name, 'Could not change spelling region preference from "' + newRegionName + '" to "' + regionName + '".');
+  } else {
+    this.webifi.output(this.name, 'Spelling region preference set to: ' + newRegionName + '.');
+  }
+}
+
 WordsWebifi.prototype.handleAnagrams = function(fodder) {
   if (!fodder) {
     return;
@@ -256,6 +278,8 @@ WordsWebifi.prototype.handler = function(input, words, commandName,
   const remaining = words.slice(numMatchedWords).join(' ');
   if (commandName == 'pattern') {
     this.handlePattern(remaining);
+  } else if (commandName == 'region') {
+    this.handleRegion(remaining);
   } else if (commandName == 'anagrams') {
     this.handleAnagrams(remaining);
   } else if (commandName == 'homophones') {